fix method delete

+++ src/core/controllers/patientController.ts
@@ -43,7 +43,13 @@ const postPatient = async (request: Request, response: Response) => {
 
 const deletePatient = async (request: Request, response: Response) => {
   try {
-    const patientDe = await patientService.deletePatient(request.params.id);
+    const isPatient = await patientService.findPatient(request.params.id);
+    if (!isPatient) {
+      return response
+        .status(404)
+        .json({ statusCode: 404, message: "Patient Not Found" });
+    }
+    await patientService.deletePatient(request.params.id);
     return response.status(201).send("The patient was deleted successfully");
   } catch (error: any) {
     return response.status(400).json({ statusCode: 400, message: `${error}` });
